First commit of reading in file
It now reads in the file. Does not read in the coefficients yet, but does get beginning of file.

Also switched most arrays to vectors or vectors of vectors, but that is not completely overhauled yet. Currently, the code is not actually functional, and most of the problem areas are due to switching to vectors and have been commented out.