Third commit of reading in file
Literally almost done with the whole shebang. There are only two minor things left to do for the code itself (at this time). Added in the test file from this iteration, to see how it looks. Discrepancies between this and the example file are because there was actually an error with the sigma values in the old version, and the new one has a narrower sigma